// src/generated/models/Crb32_casesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crb32_casesescrb32_stagename = {
  825550000: 'ابتدائي',
  825550001: 'استئناف',
  825550002: 'نقض',
  825550003: 'تنفيذ',
  825550004: 'التماس'
} as const;
export type Crb32_casesescrb32_stagename = keyof typeof Crb32_casesescrb32_stagename;
export const Crb32_casesescrb32_ty = {
  825550000: 'مدني',
  825550001: 'جنايات',
  825550002: 'عمال',
  825550003: 'تعويضات',
  825550004: 'جنح',
  825550005: 'مكتب عمل',
  825550006: 'تجاري'
} as const;
export type Crb32_casesescrb32_ty = keyof typeof Crb32_casesescrb32_ty;
export const Crb32_casesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crb32_casesesstatecode = keyof typeof Crb32_casesesstatecode;
export const Crb32_casesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crb32_casesesstatuscode = keyof typeof Crb32_casesesstatuscode;

export interface Crb32_casesesBase {
  crb32_casesid: string;
  "crb32_CurrentStage@odata.bind"?: string;
  crb32_description?: string;
  crb32_from?: string;
  crb32_link?: string;
  crb32_newcolumn?: string;
  "crb32_Responsible@odata.bind": string;
  "crb32_SecondResponsible@odata.bind"?: string;
  crb32_to?: string;
  crb32_ty?: Crb32_casesescrb32_ty;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Crb32_casesesstatecode;
  statuscode?: Crb32_casesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crb32_caseses extends Crb32_casesesBase {
  crb32_currentstagename?: string;
  crb32_responsiblename?: string;
  crb32_responsibleyominame?: string;
  crb32_secondresponsiblename?: string;
  crb32_secondresponsibleyominame?: string;
  crb32_stagename?: Crb32_casesescrb32_stagename;
  crb32_stagenamename?: string;
  crb32_tyname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  crb32_currentstage?: object;
  _crb32_currentstage_value?: string;
  crb32_responsible?: object;
  _crb32_responsible_value?: string;
  crb32_secondresponsible?: object;
  _crb32_secondresponsible_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}