// src/generated/models/Crb32_stagesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crb32_stagesescrb32_stagename = {
  825550000: 'ابتدائي',
  825550001: 'استئناف',
  825550002: 'نقض',
  825550003: 'تنفيذ',
  825550004: 'التماس'
} as const;
export type Crb32_stagesescrb32_stagename = keyof typeof Crb32_stagesescrb32_stagename;
export const Crb32_stagesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crb32_stagesesstatecode = keyof typeof Crb32_stagesesstatecode;
export const Crb32_stagesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crb32_stagesesstatuscode = keyof typeof Crb32_stagesesstatuscode;

export interface Crb32_stagesesBase {
  crb32_autonum?: string;
  "crb32_Case@odata.bind"?: string;
  crb32_claimanttext?: string;
  crb32_defendanttext?: string;
  crb32_description?: string;
  crb32_jungledistrict?: number;
  crb32_newcolumn: string;
  crb32_number?: number;
  crb32_stagename?: Crb32_stagesescrb32_stagename;
  crb32_stagesid: string;
  crb32_stageyear?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Crb32_stagesesstatecode;
  statuscode?: Crb32_stagesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crb32_stageses extends Crb32_stagesesBase {
  crb32_casename?: string;
  crb32_from?: string;
  crb32_responsible?: string;
  crb32_stagenamename?: string;
  crb32_to?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  crb32_case?: object;
  _crb32_case_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}