// src/generated/models/Crb32_updatesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crb32_updatesescrb32_stagenamecalendar = {
  825550000: 'ابتدائي',
  825550001: 'استئناف',
  825550002: 'نقض',
  825550003: 'تنفيذ',
  825550004: 'التماس'
} as const;
export type Crb32_updatesescrb32_stagenamecalendar = keyof typeof Crb32_updatesescrb32_stagenamecalendar;
export const Crb32_updatesescrb32_updatetype = {
  825550000: 'جلسة',
  825550001: 'تأجيل',
  825550002: 'حكم تمهيدي',
  825550003: 'حكم',
  825550004: 'طعن',
  825550005: 'الموقف القانوني'
} as const;
export type Crb32_updatesescrb32_updatetype = keyof typeof Crb32_updatesescrb32_updatetype;
export const Crb32_updatesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crb32_updatesesstatecode = keyof typeof Crb32_updatesesstatecode;
export const Crb32_updatesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crb32_updatesesstatuscode = keyof typeof Crb32_updatesesstatuscode;

export interface Crb32_updatesesBase {
  "crb32_Cases@odata.bind"?: string;
  crb32_currentdate?: string;
  crb32_date?: string;
  crb32_description?: string;
  crb32_documentsprovided?: string;
  crb32_emailsent?: boolean;
  crb32_newcolumn?: string;
  "crb32_Stage@odata.bind"?: string;
  crb32_updatesid: string;
  crb32_updatetype?: Crb32_updatesescrb32_updatetype;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Crb32_updatesesstatecode;
  statuscode?: Crb32_updatesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crb32_updateses extends Crb32_updatesesBase {
  crb32_casesname?: string;
  crb32_emailsentname?: string;
  crb32_recive?: string;
  crb32_secondrecive?: string;
  crb32_stagename?: string;
  crb32_stagenamecalendar?: Crb32_updatesescrb32_stagenamecalendar;
  crb32_stagenamecalendarname?: string;
  crb32_updatetypename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  crb32_cases?: object;
  _crb32_cases_value?: string;
  crb32_stage?: object;
  _crb32_stage_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}